requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
click==7.1.2
colorama==0.4.4
decorator==4.4.2
dictalchemy==0.1.2.7
fastapi==0.61.2
Flask==1.1.2
Flask-json-schema==0.0.5
Flask-JsonSchema==0.1.1
flask-jsonschema-validator==0.0.4
Flask-SQLAlchemy==2.4.4
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pydantic==1.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.2
pytest-html==3.0.0
pytest-metadata==1.10.0
PyYAML==5.3.1
self==0.0.0
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
toml==0.10.2
Werkzeug==1.0.1
